requirements.txt
streamlit==1.32.0
snowflake-connector-python==3.7.0
pandas==2.2.0
numpy==1.26.3
plotly==5.19.0
folium==0.15.1
streamlit-folium==0.18.0
altair==5.2.0
seaborn==0.13.2
matplotlib==3.8.2
scikit-learn==1.4.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
geopy==2.4.1
wordcloud==1.9.3
Pillow==10.2.0